Implemented Interface and Chnaged abstract class to Conctrete class | Mohit
