+++ Modals/CarParkingSpot.java
@@ -1,6 +1,6 @@
 package Modals;
 
-public class CarParkingSpot extends ParkingSpot{
+public class CarParkingSpot extends ParkingSpotClass{
     private static final String parkingType = "Car Parking";
     private static final int ratePerHour = 30;
