+++ Modals/Truck.java
@@ -1,5 +1,5 @@
 package Modals;
-public class Truck extends Vehicle {
+public class Truck extends VehicleClass {
     private static final String type = "TRUCK";
     private static final int ratePerHours = 50;
     private long licencePlateNumber;
