+++ Modals/TruckParkingSpot.java
@@ -1,6 +1,6 @@
 package Modals;
 
-public class TruckParkingSpot extends ParkingSpot {
+public class TruckParkingSpot extends ParkingSpotClass {
     private static final String parkingType = "Truck Parking";
     private static final int ratePerHour = 50;
